Move the org.jboss.logmanager.config package to org.wildfly.core.logmanager.config.
