+++ logging/src/main/java/org/jboss/as/logging/LoggingOperations.java
@@ -26,7 +26,7 @@ import org.jboss.as.logging.logmanager.ConfigurationPersistence;
 import org.jboss.as.server.ServerEnvironment;
 import org.jboss.dmr.ModelNode;
 import org.jboss.logmanager.LogContext;
-import org.jboss.logmanager.config.LogContextConfiguration;
+import org.wildfly.core.logmanager.config.LogContextConfiguration;
 import org.wildfly.security.manager.WildFlySecurityManager;
 
 /**
