+++ logging/src/main/java/org/jboss/as/logging/PropertyAttributeDefinition.java
@@ -12,7 +12,7 @@ import org.jboss.as.controller.SimpleAttributeDefinition;
 import org.jboss.as.logging.resolvers.ModelNodeResolver;
 import org.jboss.dmr.ModelNode;
 import org.jboss.dmr.ModelType;
-import org.jboss.logmanager.config.PropertyConfigurable;
+import org.wildfly.core.logmanager.config.PropertyConfigurable;
 
 /**
  * Defines an attribute with a property name.
