+++ logging/src/main/java/org/jboss/as/logging/deployments/LoggingDeploymentResourceProcessor.java
@@ -12,7 +12,7 @@ import org.jboss.as.server.deployment.DeploymentPhaseContext;
 import org.jboss.as.server.deployment.DeploymentResourceSupport;
 import org.jboss.as.server.deployment.DeploymentUnit;
 import org.jboss.as.server.deployment.DeploymentUnitProcessor;
-import org.jboss.logmanager.config.LogContextConfiguration;
+import org.wildfly.core.logmanager.config.LogContextConfiguration;
 import org.jboss.modules.Module;
 import org.wildfly.security.manager.WildFlySecurityManager;
