+++ logmanager/src/main/java/org/wildfly/core/logmanager/config/AbstractBasicConfiguration.java
@@ -17,7 +17,7 @@
  * limitations under the License.
  */
 
-package org.jboss.logmanager.config;
+package org.wildfly.core.logmanager.config;
 
 import java.util.Map;
