+++ logmanager/src/main/java/org/wildfly/core/logmanager/config/ConfigAction.java
@@ -1,7 +1,7 @@
 /*
  * JBoss, Home of Professional Open Source.
  *
- * Copyright 2014 Red Hat, Inc., and individual contributors
+ * Copyright 2023 Red Hat, Inc., and individual contributors
  * as indicated by the @author tags.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -17,14 +17,23 @@
  * limitations under the License.
  */
 
-package org.jboss.logmanager.config;
+package org.wildfly.core.logmanager.config;
 
-import java.util.logging.Formatter;
+import org.jboss.logmanager.configuration.ConfigurationResource;
 
 /**
- * A configuration for a logger formatter.
- *
  * @author <a href="mailto:david.lloyd@redhat.com">David M. Lloyd</a>
+ * @author <a href="mailto:jperkins@redhat.com">James R. Perkins</a>
  */
-public interface FormatterConfiguration extends NamedConfigurable, ObjectConfigurable<Formatter>, PropertyConfigurable {
+interface ConfigAction<T> {
+    T validate() throws IllegalArgumentException;
+    default T validate(ConfigurationResource<T> resource) {
+        return resource.get();
+    }
+
+    void applyPreCreate(T param);
+
+    void applyPostCreate(T param);
+
+    void rollback();
 }
