+++ logmanager/src/main/java/org/wildfly/core/logmanager/config/FilterConfigurationImpl.java
@@ -1,7 +1,7 @@
 /*
  * JBoss, Home of Professional Open Source.
  *
- * Copyright 2014 Red Hat, Inc., and individual contributors
+ * Copyright 2023 Red Hat, Inc., and individual contributors
  * as indicated by the @author tags.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -17,7 +17,7 @@
  * limitations under the License.
  */
 
-package org.jboss.logmanager.config;
+package org.wildfly.core.logmanager.config;
 
 import java.util.logging.Filter;
